+++ R/SDistribution_Arcsine.R
@@ -176,14 +176,6 @@ Arcsine <- R6Class("Arcsine",
       self$quantile(runif(n))
     },
 
-    # getRefParams
-    .getRefParams = function(paramlst) {
-      lst <- list()
-      if (!is.null(paramlst$lower)) lst <- c(lst, list(lower = paramlst$lower))
-      if (!is.null(paramlst$upper)) lst <- c(lst, list(upper = paramlst$upper))
-      return(lst)
-    },
-
     # traits
     .traits = list(valueSupport = "continuous", variateForm = "univariate")
   )