+++ R/SDistribution_ChiSquaredNoncentral.R
@@ -207,14 +207,6 @@ ChiSquaredNoncentral <- R6Class("ChiSquaredNoncentral",
       )
     },
 
-    # getRefParams
-    .getRefParams = function(paramlst) {
-      lst <- list()
-      if (!is.null(paramlst$df)) lst <- c(lst, list(df = paramlst$df))
-      if (!is.null(paramlst$location)) lst <- c(lst, list(location = paramlst$location))
-      return(lst)
-    },
-
     # traits
     .traits = list(valueSupport = "continuous", variateForm = "univariate")
   )